NL2SQL/pyproject.toml
[project]
name = "nl2sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "langgraph>=1.0.5",
    "networkx>=3.6.1",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "pydantic-ai>=1.39.0",
    "pyodbc>=5.3.0",
    "streamlit>=1.52.2",
    "tabulate>=0.9.0",
    "uvicorn>=0.40.0",
]
